Use smartcols for search output
